# Conflicts:
#	build.gradle
#	src/main/java/sch/cqre/api/config/AppConfig.java
#	src/main/java/sch/cqre/api/domain/BookEntity.java
#	src/main/java/sch/cqre/api/domain/CommentEntity.java
#	src/main/java/sch/cqre/api/domain/PostEntity.java
#	src/main/java/sch/cqre/api/domain/RentalEntity.java
#	src/main/java/sch/cqre/api/domain/RentalEntityPK.java
#	src/main/java/sch/cqre/api/domain/SupplyEntity.java
#	src/main/java/sch/cqre/api/exception/CustomException.java
#	src/main/java/sch/cqre/api/exception/ErrorCode.java
#	src/main/java/sch/cqre/api/exception/ErrorResponse.java
#	src/main/java/sch/cqre/api/handler/GlobalExceptionHandler.java

+++ src/main/java/sch/cqre/api/domain/RentalEntityPK.java
@@ -12,11 +12,11 @@ import java.io.Serializable;
 public class RentalEntityPK implements Serializable {
     @Column(name = "user_id")
     @Id
-    private Integer userId;
+    private Long userId;
 
     @Column(name = "book_id")
     @Id
-    private Integer bookId;
+    private Long bookId;
 
     @Override
     public boolean equals(Object o) {
